Fix B-spline cache locality benchmark issues
Address three issues identified in review:

1. Add missing #include <algorithm> for std::sort
   - benchmark.cc:57 uses std::sort but didn't include <algorithm>
   - Would fail to compile when added to Bazel target

2. Move benchmark to benchmarks/ and add to BUILD.bazel
   - Moved from tests/ to benchmarks/ (correct location)
   - Added cc_binary target with tags = ["benchmark", "manual"]
   - Follows pattern of other benchmarks (component_performance, etc.)

3. Implement actual old vs new order comparison
   - Created BSplineFitterOldOrder class to replicate pre-optimization axis order
   - Benchmark now runs BOTH old (m → τ → σ → r) and new (r → σ → τ → m) orders
   - Outputs side-by-side timings and computed speedup
   - Validates optimization claim with empirical measurements

Now the benchmark can be built with:
  bazel build //benchmarks:bspline_cache_locality

And run with:
  bazel run //benchmarks:bspline_cache_locality

This provides concrete evidence of the cache locality improvement.
