--- tests/bspline_cache_locality_benchmark.cc
@@ -1,153 +1,0 @@
-/**
- * @file bspline_cache_locality_benchmark.cc
- * @brief Benchmark to demonstrate cache locality improvement from axis reordering
- *
- * This benchmark measures the performance impact of processing axes in order
- * of increasing memory stride (r → σ → τ → m) vs. the old order (m → τ → σ → r).
- *
- * Expected results:
- * - Old order: m-axis has stride Nt*Nv*Nr (~48KB for 50×30×20×10), poor cache locality
- * - New order: r-axis has stride 1 (contiguous), excellent cache locality
- * - Speedup: Hardware-dependent, typically 1.5-3x for large grids
- */
-
-#include "src/bspline_fitter_4d_separable.hpp"
-#include <chrono>
-#include <iostream>
-#include <vector>
-#include <cmath>
-#include <iomanip>
-
-using namespace mango;
-
-/// Helper: Create linearly spaced grid
-std::vector<double> linspace(double start, double end, int n) {
-    std::vector<double> result(n);
-    for (int i = 0; i < n; ++i) {
-        result[i] = start + (end - start) * i / (n - 1);
-    }
-    return result;
-}
-
-/// Test function: smooth separable function
-double test_function(double m, double t, double v, double r) {
-    return std::exp(-m*m) * std::exp(-t) * std::sin(v * 3.14159) * (1.0 + r);
-}
-
-/// Benchmark helper
-template<typename Func>
-double benchmark(Func&& f, int iterations = 3) {
-    std::vector<double> times;
-    times.reserve(iterations);
-
-    for (int i = 0; i < iterations; ++i) {
-        auto start = std::chrono::high_resolution_clock::now();
-        f();
-        auto end = std::chrono::high_resolution_clock::now();
-
-        double elapsed = std::chrono::duration<double, std::milli>(end - start).count();
-        times.push_back(elapsed);
-    }
-
-    // Return median time
-    std::sort(times.begin(), times.end());
-    return times[times.size() / 2];
-}
-
-int main() {
-    std::cout << "=================================================================\n";
-    std::cout << "B-Spline Cache Locality Benchmark\n";
-    std::cout << "=================================================================\n\n";
-
-    // Test different grid sizes
-    struct GridConfig {
-        int Nm, Nt, Nv, Nr;
-        const char* description;
-    };
-
-    std::vector<GridConfig> configs = {
-        {20, 15, 10, 8,  "Small  (20×15×10×8)"},
-        {50, 30, 20, 10, "Medium (50×30×20×10)"},
-        {80, 40, 25, 15, "Large  (80×40×25×15)"},
-    };
-
-    for (const auto& config : configs) {
-        std::cout << "Grid: " << config.description << "\n";
-        std::cout << "  Total points: " << (config.Nm * config.Nt * config.Nv * config.Nr) << "\n";
-
-        // Compute strides
-        int stride_r = 1;
-        int stride_sigma = config.Nr;
-        int stride_tau = config.Nv * config.Nr;
-        int stride_m = config.Nt * config.Nv * config.Nr;
-
-        std::cout << "  Memory strides:\n";
-        std::cout << "    r-axis: " << stride_r << " doubles = " << (stride_r * 8) << " bytes\n";
-        std::cout << "    σ-axis: " << stride_sigma << " doubles = " << (stride_sigma * 8) << " bytes\n";
-        std::cout << "    τ-axis: " << stride_tau << " doubles = " << (stride_tau * 8) << " bytes\n";
-        std::cout << "    m-axis: " << stride_m << " doubles = " << (stride_m * 8) << " bytes";
-
-        if (stride_m * 8 > 32768) {
-            std::cout << " (exceeds L1 cache!)";
-        }
-        std::cout << "\n\n";
-
-        // Create grids
-        auto m_grid = linspace(0.7, 1.3, config.Nm);
-        auto t_grid = linspace(0.1, 2.0, config.Nt);
-        auto v_grid = linspace(0.1, 0.8, config.Nv);
-        auto r_grid = linspace(0.0, 0.1, config.Nr);
-
-        // Generate test data
-        std::vector<double> values(config.Nm * config.Nt * config.Nv * config.Nr);
-        for (int i = 0; i < config.Nm; ++i) {
-            for (int j = 0; j < config.Nt; ++j) {
-                for (int k = 0; k < config.Nv; ++k) {
-                    for (int l = 0; l < config.Nr; ++l) {
-                        size_t idx = ((i * config.Nt + j) * config.Nv + k) * config.Nr + l;
-                        values[idx] = test_function(m_grid[i], t_grid[j], v_grid[k], r_grid[l]);
-                    }
-                }
-            }
-        }
-
-        // Benchmark fitting
-        BSplineFitter4DSeparable fitter(m_grid, t_grid, v_grid, r_grid);
-
-        auto fit_time = benchmark([&]() {
-            auto result = fitter.fit(values, 1e-6);
-            if (!result.success) {
-                std::cerr << "Fit failed: " << result.error_message << "\n";
-            }
-        }, 5);  // 5 iterations for better statistics
-
-        std::cout << "  Fit time: " << std::fixed << std::setprecision(2) << fit_time << " ms\n";
-
-        // Compute operations per second (approximate)
-        // Each axis fit: Nt*Nv*Nr slices for m-axis, etc.
-        size_t total_tridiag_solves =
-            config.Nt * config.Nv * config.Nr +  // m-axis
-            config.Nm * config.Nv * config.Nr +  // τ-axis
-            config.Nm * config.Nt * config.Nr +  // σ-axis
-            config.Nm * config.Nt * config.Nv;   // r-axis
-
-        double solves_per_ms = total_tridiag_solves / fit_time;
-        std::cout << "  Throughput: " << std::fixed << std::setprecision(0)
-                  << solves_per_ms << " tridiagonal solves/ms\n";
-
-        std::cout << "\n";
-    }
-
-    std::cout << "=================================================================\n";
-    std::cout << "Optimization Notes:\n";
-    std::cout << "=================================================================\n";
-    std::cout << "The cache-optimized axis order (r → σ → τ → m) processes the\n";
-    std::cout << "fastest-varying dimensions first, keeping the working set hot in\n";
-    std::cout << "cache during early passes.\n\n";
-    std::cout << "For the old order (m → τ → σ → r), the m-axis pass would access\n";
-    std::cout << "memory with stride Nt*Nv*Nr, often exceeding L1 cache size.\n\n";
-    std::cout << "Expected speedup: 1.5-3x for large grids (hardware-dependent)\n";
-    std::cout << "=================================================================\n";
-
-    return 0;
-}
